+++ packages/web/src/lib/links.ts
@@ -2,7 +2,7 @@ export const REPO_URL = "https://github.com/dishwasher-detergent/mault";
 export const MODEL_URL = `${REPO_URL}/tree/master/3d%20model`;
 export const DISCORD_URL = "https://discord.gg/fYvw5PcvGg";
 // Our fork (this project) and the original project we forked from.
-export const FORK_REPO_URL = "https://github.com/4holdingdirectlimited/mault";
+export const FORK_REPO_URL = "https://github.com/4holdingdirectlimited/DeckSift";
 export const FORK_ISSUES_URL = `${FORK_REPO_URL}/issues/new`;
 export const UPSTREAM_URL = "https://mault.xyz";
 export const UPSTREAM_REPO_URL = REPO_URL;